Moved exceptions to separate Exceptions.cs file
