+++ Editor/Components/UiPanelConfigEntryDrawer.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Threading;
 using UnityEditor;
 using UnityEngine;
-using GuiToolkit.AssetHandling;
 using Object = UnityEngine.Object;
 
 namespace GuiToolkit.AssetHandling
@@ -19,11 +18,12 @@ namespace GuiToolkit.AssetHandling
 		// warnings
 		private string m_lastWarning;
 		private double m_warningUntil; // EditorApplication.timeSinceStartup deadline
+		private static IAssetProvider[] s_assetProviders;
 
 		public override void OnGUI( Rect _position, SerializedProperty _property, GUIContent _label )
 		{
 			var typeProp = _property.FindPropertyRelative("Type");
-			var idProp   = _property.FindPropertyRelative("PanelId");
+			var idProp = _property.FindPropertyRelative("PanelId");
 
 			EditorGUI.BeginProperty(_position, _label, _property);
 
@@ -32,21 +32,25 @@ namespace GuiToolkit.AssetHandling
 			EditorGUI.LabelField(row, new GUIContent(string.IsNullOrEmpty(typeProp.stringValue) ? "<Type not set>" : typeProp.stringValue));
 			row.y += kLine + kPadY;
 
+			if (s_assetProviders == null)
+			{
+				try
+				{
+					s_assetProviders = AssetManager.AssetProviders;
+				}
+				catch (Exception ex)
+				{
+					
+					DrawWarning(ref row, _position, "AssetManager not initialized: " + ex.Message);
+					EditorGUI.EndProperty();
+					return;
+				}
+
+			}
 			// Get providers from manager (safe guard)
-			IAssetProvider[] providers = Array.Empty<IAssetProvider>();
-			try
-			{
-				providers = AssetManager.AssetProviders;
-			}
-			catch (Exception ex)
-			{
-				DrawWarning(ref row, _position, "AssetManager not initialized: " + ex.Message);
-				EditorGUI.EndProperty();
-				return;
-			}
 
 			// Draw one input per provider
-			foreach (var provider in providers)
+			foreach (var provider in s_assetProviders)
 			{
 				DrawProviderRow(ref row, _position, provider, typeProp, idProp);
 			}
@@ -90,7 +94,7 @@ namespace GuiToolkit.AssetHandling
 		{
 			// Label (provider name + ResName)
 			float labelWidth = Mathf.Min(160f, _bounds.width * 0.35f);
-			var left  = new Rect(_row.x, _row.y, labelWidth, kLine);
+			var left = new Rect(_row.x, _row.y, labelWidth, kLine);
 			var right = new Rect(_row.x + labelWidth + 4f, _row.y, _bounds.width - labelWidth - 4f, kLine);
 
 			string providerLabel = string.IsNullOrEmpty(_provider.ResName)
